Fixed viewport 16x9

+++ core/src/main/java/com/IONA/TowerDefense/model/Waves.java
@@ -9,7 +9,7 @@ public class Waves {
     public static class Enemy {
         public String type;
         //public String difficulty;
-        public int delay;
+        public double delay;
     }
 
     public static class Wave {
